# Only do realistic request errors when we need it