--- a/spec/support/error_responses_helper.rb
+++ b/spec/support/error_responses_helper.rb
@@ -1,0 +1,23 @@
+# https://github.com/eliotsykes/rails-testing-toolbox/blob/master/error_responses.rb
+module ErrorResponses
+  # See: https://github.com/rails/rails/pull/11289#issuecomment-118612393
+  def respond_without_detailed_exceptions
+    env_config = Rails.application.env_config
+    original_show_exceptions = env_config["action_dispatch.show_exceptions"]
+    original_show_detailed_exceptions = env_config["action_dispatch.show_detailed_exceptions"]
+    env_config["action_dispatch.show_exceptions"] = true
+    env_config["action_dispatch.show_detailed_exceptions"] = false
+    yield
+  ensure
+    env_config["action_dispatch.show_exceptions"] = original_show_exceptions
+    env_config["action_dispatch.show_detailed_exceptions"] = original_show_detailed_exceptions
+  end
+end
+
+RSpec.configure do |config|
+  config.include ErrorResponses
+
+  config.around(realistic_error_responses: true) do |example|
+    respond_without_detailed_exceptions(&example)
+  end
+end