@benthorner 2020-03-12
@kevindew just to document a few things from our discussion. We're trying to solve a couple of problems:

   - It's laborious/confusing to debug errors that happen during a request.
   - Tests involving requests take longer to run than they need to.

I'm not set on this particular PR, although I would like a solution to be consistent across all tests (involving requests). Some of the alternatives we discussed:

   - Changing `config.consider_all_requests_local = true` to `false`, which still achieves the performance gain. This would mean all response bodies show render the page associated with their response code, which means our tests are more consistent with the production behaviour. It would mean we have to use `test.log` to debug anything.

   - Modifying the approach in this PR to (only) apply to all request tests. I'm struggling to see the advantage with this one, over what we have now: we lose the performance gains, and our test behaviour becomes fragmented. I think it would be better to stay as-is.

Hope that helps. No rush to get this in or anything - it just seemed like a good time to raise it. We could also move it to a pain card if the way forward is unclear.