@kevindew 2020-03-12
Thanks for raising this @benthorner. I'll give this some more attention when I have some spare moments.