@benthorner 2020-04-07
Closing this for now, as I don't think we'll have time to discuss it much further. Although it would lead to improved performance in running the tests, the approach suggested here is still a fragmented one, in that the debugging process for 'tests with realistic errors' is different. This behaviour does cause some confusion in general, so we should document the debug method.

Closing in favour of https://github.com/alphagov/content-publisher/pull/1991.